+++ src/Exceptions/AssetNotAvailableException.php
@@ -1,0 +1,11 @@
+<?php
+namespace Netcoins\Exceptions;
+
+/**
+ * Exception to be thrown when a client method is provided with an asset that is not found on Netcoins
+ *
+ * @author Simon Willan <swillan@gonetcoins.com>
+ */
+class AssetNotAvailableException extends \Exception
+{
+}